clean up optimization allocations and loops with tests for runtime and float precision change

+++ layernorm-systems/optimized/backward.cpp
@@ -7,14 +7,14 @@
 using namespace std;
 
 // function definition for performing backward pass on output and cache via layernorm
-backwardOutput backwardPassLayerNorm(torch::Tensor dout, vector<torch::Tensor> cache, double epsilon) {
+backwardOutput backwardPassLayerNorm(torch::Tensor dout, vector<torch::Tensor> cache, float epsilon) {
 
     // extract intermediate values from forward pass cache (also guarantee row-major)
-    torch::Tensor gamma = cache[0].contiguous();
-    torch::Tensor xhat = cache[1].contiguous();
-    torch::Tensor xmu = cache[2].contiguous();
-    torch::Tensor sqrtvar = cache[3].contiguous();
-    torch::Tensor ivar = cache[4].contiguous();
+    torch::Tensor gamma = cache[0].contiguous();  // 1-D
+    torch::Tensor xhat = cache[1].contiguous();  // 2-D
+    torch::Tensor xmu = cache[2].contiguous();  // 2-D
+    torch::Tensor sqrtvar = cache[3].contiguous();  // 1-D
+    torch::Tensor ivar = cache[4].contiguous();  // 1-D
 
     // validate forward pass output with backward pass input tensor (contract)
     TORCH_CHECK(dout.sizes() == xhat.sizes(), "Input tensor does not match the correct element size.");  // element-wise ops
@@ -32,31 +32,31 @@ backwardOutput backwardPassLayerNorm(torch::Tensor dout, vector<torch::Tensor> c
     torch::Tensor dbeta = torch::zeros({dims}, gamma.options());  // accumulator (beta)
 
     // initiate data pointers for output tensors (element-wise ops)
-    double *ptr_dx = dx.data_ptr<double>();  // 2-D (N, D)
-    double *ptr_dgamma = dgamma.data_ptr<double>();  // 1-D (D,)
-    double *ptr_dbeta = dbeta.data_ptr<double>();  // 1-D (D,)
+    float *ptr_dx = dx.data_ptr<float>();  // 2-D (N, D)
+    float *ptr_dgamma = dgamma.data_ptr<float>();  // 1-D (D,)
+    float *ptr_dbeta = dbeta.data_ptr<float>();  // 1-D (D,)
     
     // initiate data pointers for cache tensors (element-wise ops)
-    double *ptr_dout = dout.data_ptr<double>();  // 2-D (N, D)
-    double *ptr_gamma = gamma.data_ptr<double>();  // 1-D (D,)
-    double *ptr_xhat = xhat.data_ptr<double>();  // 2-D (N, D)
-    double *ptr_xmu = xmu.data_ptr<double>();  // 2-D (N, D)
-    double *ptr_sqrtvar = sqrtvar.data_ptr<double>();  // 1-D (N,)
-    double *ptr_ivar = ivar.data_ptr<double>();  // 1-D (N,)
+    float *ptr_dout = dout.data_ptr<float>();  // 2-D (N, D)
+    float *ptr_gamma = gamma.data_ptr<float>();  // 1-D (D,)
+    float *ptr_xhat = xhat.data_ptr<float>();  // 2-D (N, D)
+    float *ptr_xmu = xmu.data_ptr<float>();  // 2-D (N, D)
+    float *ptr_sqrtvar = sqrtvar.data_ptr<float>();  // 1-D (N,)
+    float *ptr_ivar = ivar.data_ptr<float>();  // 1-D (N,)
 
     // create intermediate ops tensors (temp)
     torch::Tensor dvar = torch::empty({n}, xhat.options());  // 1-D but xhat parameters
     torch::Tensor dxmu2 = torch::empty_like(xhat);
 
     // initiate data pointers for intermediate ops tensors
-    double *ptr_dvar = dvar.data_ptr<double>();  // 1-D (N,)
-    double *ptr_dxmu2 = dxmu2.data_ptr<double>();  // 2-D (N, D)
+    float *ptr_dvar = dvar.data_ptr<float>();  // 1-D (N,)
+    float *ptr_dxmu2 = dxmu2.data_ptr<float>();  // 2-D (N, D)
 
     // iterate through rows/groups
     for (int i = 0; i < n; i++) {
 
         // linear scaling for learnable parameters (gradients)
-        double divar_sum = 0.0;
+        float divar_sum = 0.0f;
         for (int j = 0; j < dims; j++) {
             ptr_dbeta[j] += ptr_dout[(i * dims) + j];  // dbeta
             ptr_dgamma[j] += (ptr_dout[(i * dims) + j] * ptr_xhat[(i * dims) + j]);  // dgamma
@@ -64,18 +64,18 @@ backwardOutput backwardPassLayerNorm(torch::Tensor dout, vector<torch::Tensor> c
         }
 
         // calculate gradient (w.r.t.) variance via chain rule
-        ptr_dvar[i] = 0.5 * (1.0 / ptr_sqrtvar[i]) * ((-1.0 / (ptr_sqrtvar[i] * ptr_sqrtvar[i])) * divar_sum);  // dvar
+        ptr_dvar[i] = 0.5f * (1.0f / ptr_sqrtvar[i]) * ((-1.0f / (ptr_sqrtvar[i] * ptr_sqrtvar[i])) * divar_sum);  // dvar
 
         // calculate gradient (w.r.t.) squared deviations via chain rule
-        double dmu_sum = 0.0;
+        float dmu_sum = 0.0f;
         for (int j = 0; j < dims; j++) {
-            ptr_dxmu2[(i * dims) + j] = 2.0 * ptr_xmu[(i * dims) + j] * ((1.0 / dims) * ptr_dvar[i]);  // dxmu2
+            ptr_dxmu2[(i * dims) + j] = 2.0f * ptr_xmu[(i * dims) + j] * ((1.0f / dims) * ptr_dvar[i]);  // dxmu2
             dmu_sum += (((ptr_dout[(i * dims) + j] * ptr_gamma[j])) * ptr_ivar[i]) + ptr_dxmu2[(i * dims) + j];  // mean accumulator
         }
 
         // calculate components of gradient (w.r.t.) mean computations
         for (int j = 0; j < dims; j++) {
-            ptr_dx[(i * dims) + j] = ((((ptr_dout[(i * dims) + j] * ptr_gamma[j])) * ptr_ivar[i]) + ptr_dxmu2[(i * dims) + j]) + ((1.0 / dims) * (-1.0 * dmu_sum));  // dx
+            ptr_dx[(i * dims) + j] = ((((ptr_dout[(i * dims) + j] * ptr_gamma[j])) * ptr_ivar[i]) + ptr_dxmu2[(i * dims) + j]) + ((1.0f / dims) * (-1.0f * dmu_sum));  // dx
         }
 
     }
